+++ test/partial-recall-test.el
@@ -484,7 +484,6 @@
 
 ;; -- Completion
 
-
 (ert-deftest pr--complete-dream ()
   (bydi-with-mock ((:mock completing-read :return "second")
                    (:ignore partial-recall--reality-owns-buffer-p)
@@ -498,6 +497,18 @@
                    (:mock partial-recall--mapped-buffers :return '("first" "second" "third"))
                    (:mock buffer-name :with bydi-rf))
     (should (string= "second" (partial-recall--complete-reality "Some prompt: ")))))
+
+(ert-deftest pr--complete-reality--inital-input ()
+  (bydi (completing-read
+         (:always partial-recall--reality-owns-buffer-p)
+         (:mock partial-recall--mapped-buffers :return '(first second third))
+         (:mock buffer-name :with bydi-rf)
+         (:mock current-buffer :return 'second))
+
+    (partial-recall--complete-reality "Some prompt: ")
+
+    (bydi-was-called-with completing-read
+      '("Some prompt: " ((first . first) (second . second) (third . third)) nil t second))))
 
 (ert-deftest pr--complete-subconscious ()
   (let* ((partial-recall--table (make-hash-table))